ArmPkg/BdsLib: Fixed the alignment of the relocated Device Tree
When the Device Tree was loaded above the 32bit address space the operation:
*RelocatedFdt = ALIGN ((UINT64)*RelocatedFdt, (UINT32)FdtAlignment);
was returning an incorrect address.

Contributed-under: TianoCore Contribution Agreement 1.0



git-svn-id: https://svn.code.sf.net/p/edk2/code/trunk/edk2@14915 6f19259b-4bc3-4df7-8a09-765794883524

+++ ArmPkg/Library/BdsLib/BdsLinuxFdt.c
@@ -264,7 +264,7 @@ RelocateFdt (
 {
   EFI_STATUS            Status;
   INTN                  Error;
-  UINT32                FdtAlignment;
+  UINT64                FdtAlignment;
 
   *RelocatedFdtSize = OriginalFdtSize + FDT_ADDITIONAL_ENTRIES_SIZE;
